Deleted faulty window listener that resulted in an irritating mobile experience.
The automatic resizing which occurs in mobile browsers caused a recurring resetting of the lesson information section.  The code was deleted, and I am seeking a new solution for a display issue which occurs when the browser window is resized small (to a mobile size), lesson info tabs are opened, and the window is made large again.  In this instance, descriptive text may be displayed overlapping other descriptive text until one of the options on the left is clicked.